Refactor structure
