+++ site/cmd/main.go
@@ -1,20 +1,9 @@
 package main
 
 import (
-	"log"
-
-	"github.com/Manas-Project-Aurora/gavna/site/cmd/api"
-	"github.com/Manas-Project-Aurora/gavna/site/config"
-	"github.com/gin-gonic/gin"
+	"github.com/Manas-Project-Aurora/gavna/site/internal/server"
 )
 
 func main() {
-	db, err := config.ConnectToDB()
-	if err != nil {
-		log.Fatalf("Базе пизда: %v", err)
-	}
-	router := gin.Default()
-	api.RegisterRoutes(router, db)
-	log.Println("Server running on 8080")
-	router.Run(":8080")
+	server.NewServer(8080).Run()
 }
