+++ site/internal/services/vacancy_service.go
@@ -2,7 +2,7 @@ package services
 
 import (
 	"github.com/Manas-Project-Aurora/gavna/internal/models"
-	"github.com/Manas-Project-Aurora/gavna/site/repository"
+	"github.com/Manas-Project-Aurora/gavna/site/internal/repository"
 )
 
 type VacancyService interface {
